+++ src/pages/Landing/Button/Button.tsx
@@ -1,5 +1,5 @@
+import Button from "@/components/Common/Button";
 import { useNavigate } from "react-router-dom";
-import Button from "../../../components/Common/Button";
 
 function Landing() {
   const navigate = useNavigate();
